# From NuPIC: Can inhibitColumnsGlobal() execute without sort() ? #354 (#365) 