--- a/src/test/java/org/numenta/nupic/util/ArrayUtilsTest.java
+++ b/src/test/java/org/numenta/nupic/util/ArrayUtilsTest.java
@@ -35,177 +35,177 @@ import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
 
 public class ArrayUtilsTest {
-    
-    @Test
-    public void testConcat() {
-        // Test happy path
-        double[] one = new double[] { 1., 2., 3. };
-        double[] two = new double[] { 4., 5., 6. };
-        double[] retVal = ArrayUtils.concat(one, two);
-        assertEquals(6, retVal.length);
-        for(int i = 0;i < retVal.length;i++) {
-            assertEquals(i + 1, retVal[i], 0);
-        }
-        
-        // Test unequal sizes
-        one = new double[] { 1., 2. };
-        retVal = ArrayUtils.concat(one, two);
-        assertEquals(5, retVal.length);
-        for(int i = 0;i < retVal.length;i++) {
-            if(i == 2) continue;
-            assertEquals(i + 1, retVal[i > 2 ? i - 1 : i], 0);
-        }
-        
-        one = new double[] { 1., 2., 3. };
-        two = new double[] { 4., 5. };
-        retVal = ArrayUtils.concat(one, two);
-        assertEquals(5, retVal.length);
-        for(int i = 0;i < retVal.length;i++) {
-            assertEquals(i + 1, retVal[i], 0);
-        }
-        
-        //Test zero length
-        one = new double[0];
-        two = new double[] { 4., 5., 6. };
-        retVal = ArrayUtils.concat(one, two);
-        assertEquals(3, retVal.length);
-        for(int i = 0;i < retVal.length;i++) {
-            assertEquals(i + 4, retVal[i], 0);
-        }
-        
-        one = new double[] { 1., 2., 3. };
-        two = new double[0];
-        retVal = ArrayUtils.concat(one, two);
-        assertEquals(3, retVal.length);
-        for(int i = 0;i < retVal.length;i++) {
-            assertEquals(i + 1, retVal[i], 0);
-        }
-        
-    }
-    
-    @Test
-    public void testInterleave() {
-        String[] f = { "0" };
-        double[] s = { 0.8 };
-         
-        // Test most simple interleave of equal length arrays
-        Object[] result = ArrayUtils.interleave(f, s);
-        assertEquals("0", result[0]);
-        assertEquals(0.8, result[1]);
-        
-        // Test simple interleave of larger array
-        f = new String[] { "0", "1" };
-        s = new double[] { 0.42, 2.5 };
-        result = ArrayUtils.interleave(f, s);
-        assertEquals("0", result[0]);
-        assertEquals(0.42, result[1]);
-        assertEquals("1", result[2]);
-        assertEquals(2.5, result[3]);
-        
-        // Test complex interleave of larger array
-        f = new String[] { "0", "1", "bob", "harry", "digit", "temperature" };
-        s = new double[] { 0.42, 2.5, .001, 1e-2, 34.0, .123 };
-        result = ArrayUtils.interleave(f, s);
-        for(int i = 0, j = 0;j < result.length;i++, j+=2) {
-            assertEquals(f[i], result[j]);
-            assertEquals(s[i], result[j + 1]);
-        }
-        
-        // Test interleave with zero length of first
-        f = new String[0];
-        s = new double[] { 0.42, 2.5 };
-        result = ArrayUtils.interleave(f, s);
-        assertEquals(0.42, result[0]);
-        assertEquals(2.5, result[1]);
-        
-        // Test interleave with zero length of second
-        f = new String[] { "0", "1" };
-        s = new double[0];
-        result = ArrayUtils.interleave(f, s);
-        assertEquals("0", result[0]);
-        assertEquals("1", result[1]);
-        
-        // Test complex unequal length: left side smaller
-        f = new String[] { "0", "1", "bob" };
-        s = new double[] { 0.42, 2.5, .001, 1e-2, 34.0, .123 };
-        result = ArrayUtils.interleave(f, s);
-        assertEquals("0", result[0]);
-        assertEquals(0.42, result[1]);
-        assertEquals("1", result[2]);
-        assertEquals(2.5, result[3]);
-        assertEquals("bob", result[4]);
-        assertEquals(.001, result[5]);
-        assertEquals(1e-2, result[6]);
-        assertEquals(34.0, result[7]);
-        assertEquals(.123, result[8]);
-        
-        // Test complex unequal length: right side smaller
-        f = new String[] { "0", "1", "bob", "harry", "digit", "temperature" };
-        s = new double[] { 0.42, 2.5, .001 };
-        result = ArrayUtils.interleave(f, s);
-        assertEquals("0", result[0]);
-        assertEquals(0.42, result[1]);
-        assertEquals("1", result[2]);
-        assertEquals(2.5, result[3]);
-        assertEquals("bob", result[4]);
-        assertEquals(.001, result[5]);
-        assertEquals("harry", result[6]);
-        assertEquals("digit", result[7]);
-        assertEquals("temperature", result[8]);
-        
-        // Negative testing
-        try {
-            f = null;
-            s = new double[] { 0.42, 2.5, .001 };
-            result = ArrayUtils.interleave(f, s);
-            fail();
-        }catch(Exception e) {
-            assertEquals(NullPointerException.class, e.getClass());
-        }
-    }
-    
-    @Test
-    public void testIn1d() {
-        int[] ar1 = { 0, 1, 5, 9, 3, 1000 };
-        int[] ar2 = Arrays.copyOf(ar1, ar1.length);
-        assertTrue(Arrays.equals(ar1, ar2));
-        int[] retVal = ArrayUtils.in1d(ar1, ar2);
-        assertTrue(Arrays.equals(ar1, ArrayUtils.reverse(retVal)));
-        
-        ar1 = new int[] { 0, 2, 1000 };
-        int[] expected = { 0, 1000 };
-        assertTrue(Arrays.equals(expected, ArrayUtils.reverse(ArrayUtils.in1d(ar1, ar2))));
-        
-        ar1 = new int[] { 2, 6, 4 };
-        expected = new int[0];
-        assertTrue(Arrays.equals(expected, ArrayUtils.in1d(ar1, ar2)));
-        
-        // Test none in the second
-        assertTrue(Arrays.equals(expected, ArrayUtils.in1d(ar1, expected)));
-        // Test none in both
-        assertTrue(Arrays.equals(expected, ArrayUtils.in1d(expected, expected)));
-    }
 
-    @Test
-    public void testRecursiveCoordinatesAssemble() throws InterruptedException {
-        /*Create huge 5 dimensional matrix*/
-        int dimSize = 14, dimNumber = 5;
-        int[] dimCoordinates = new int[dimSize];
-        List<int[]> dimensions = new ArrayList<int[]>();
-        for (int i = 0; i < dimNumber; i++) {
-            for (int j = 0; j < dimSize; j++) {
-                dimCoordinates[j] = j;
-            }
-            dimensions.add(dimCoordinates);
-        }
-        long startTime = System.currentTimeMillis();
-        
-        List<int[]> neighborList = ArrayUtils.dimensionsToCoordinateList(dimensions);
-        long take = System.currentTimeMillis() - startTime;
-        System.out.print("Execute in:" + take + " milliseconds");
+	@Test
+	public void testConcat() {
+		// Test happy path
+		double[] one = new double[] { 1., 2., 3. };
+		double[] two = new double[] { 4., 5., 6. };
+		double[] retVal = ArrayUtils.concat(one, two);
+		assertEquals(6, retVal.length);
+		for(int i = 0;i < retVal.length;i++) {
+			assertEquals(i + 1, retVal[i], 0);
+		}
 
-        assertEquals(neighborList.size(), 537824);
-    }
+		// Test unequal sizes
+		one = new double[] { 1., 2. };
+		retVal = ArrayUtils.concat(one, two);
+		assertEquals(5, retVal.length);
+		for(int i = 0;i < retVal.length;i++) {
+			if(i == 2) continue;
+			assertEquals(i + 1, retVal[i > 2 ? i - 1 : i], 0);
+		}
+
+		one = new double[] { 1., 2., 3. };
+		two = new double[] { 4., 5. };
+		retVal = ArrayUtils.concat(one, two);
+		assertEquals(5, retVal.length);
+		for(int i = 0;i < retVal.length;i++) {
+			assertEquals(i + 1, retVal[i], 0);
+		}
+
+		//Test zero length
+		one = new double[0];
+		two = new double[] { 4., 5., 6. };
+		retVal = ArrayUtils.concat(one, two);
+		assertEquals(3, retVal.length);
+		for(int i = 0;i < retVal.length;i++) {
+			assertEquals(i + 4, retVal[i], 0);
+		}
+
+		one = new double[] { 1., 2., 3. };
+		two = new double[0];
+		retVal = ArrayUtils.concat(one, two);
+		assertEquals(3, retVal.length);
+		for(int i = 0;i < retVal.length;i++) {
+			assertEquals(i + 1, retVal[i], 0);
+		}
+
+	}
+
+	@Test
+	public void testInterleave() {
+		String[] f = { "0" };
+		double[] s = { 0.8 };
+
+		// Test most simple interleave of equal length arrays
+		Object[] result = ArrayUtils.interleave(f, s);
+		assertEquals("0", result[0]);
+		assertEquals(0.8, result[1]);
+
+		// Test simple interleave of larger array
+		f = new String[] { "0", "1" };
+		s = new double[] { 0.42, 2.5 };
+		result = ArrayUtils.interleave(f, s);
+		assertEquals("0", result[0]);
+		assertEquals(0.42, result[1]);
+		assertEquals("1", result[2]);
+		assertEquals(2.5, result[3]);
+
+		// Test complex interleave of larger array
+		f = new String[] { "0", "1", "bob", "harry", "digit", "temperature" };
+		s = new double[] { 0.42, 2.5, .001, 1e-2, 34.0, .123 };
+		result = ArrayUtils.interleave(f, s);
+		for(int i = 0, j = 0;j < result.length;i++, j+=2) {
+			assertEquals(f[i], result[j]);
+			assertEquals(s[i], result[j + 1]);
+		}
+
+		// Test interleave with zero length of first
+		f = new String[0];
+		s = new double[] { 0.42, 2.5 };
+		result = ArrayUtils.interleave(f, s);
+		assertEquals(0.42, result[0]);
+		assertEquals(2.5, result[1]);
+
+		// Test interleave with zero length of second
+		f = new String[] { "0", "1" };
+		s = new double[0];
+		result = ArrayUtils.interleave(f, s);
+		assertEquals("0", result[0]);
+		assertEquals("1", result[1]);
+
+		// Test complex unequal length: left side smaller
+		f = new String[] { "0", "1", "bob" };
+		s = new double[] { 0.42, 2.5, .001, 1e-2, 34.0, .123 };
+		result = ArrayUtils.interleave(f, s);
+		assertEquals("0", result[0]);
+		assertEquals(0.42, result[1]);
+		assertEquals("1", result[2]);
+		assertEquals(2.5, result[3]);
+		assertEquals("bob", result[4]);
+		assertEquals(.001, result[5]);
+		assertEquals(1e-2, result[6]);
+		assertEquals(34.0, result[7]);
+		assertEquals(.123, result[8]);
+
+		// Test complex unequal length: right side smaller
+		f = new String[] { "0", "1", "bob", "harry", "digit", "temperature" };
+		s = new double[] { 0.42, 2.5, .001 };
+		result = ArrayUtils.interleave(f, s);
+		assertEquals("0", result[0]);
+		assertEquals(0.42, result[1]);
+		assertEquals("1", result[2]);
+		assertEquals(2.5, result[3]);
+		assertEquals("bob", result[4]);
+		assertEquals(.001, result[5]);
+		assertEquals("harry", result[6]);
+		assertEquals("digit", result[7]);
+		assertEquals("temperature", result[8]);
+
+		// Negative testing
+		try {
+			f = null;
+			s = new double[] { 0.42, 2.5, .001 };
+			result = ArrayUtils.interleave(f, s);
+			fail();
+		}catch(Exception e) {
+			assertEquals(NullPointerException.class, e.getClass());
+		}
+	}
+
+	@Test
+	public void testIn1d() {
+		int[] ar1 = { 0, 1, 5, 9, 3, 1000 };
+		int[] ar2 = Arrays.copyOf(ar1, ar1.length);
+		assertTrue(Arrays.equals(ar1, ar2));
+		int[] retVal = ArrayUtils.in1d(ar1, ar2);
+		assertTrue(Arrays.equals(ar1, ArrayUtils.reverse(retVal)));
+
+		ar1 = new int[] { 0, 2, 1000 };
+		int[] expected = { 0, 1000 };
+		assertTrue(Arrays.equals(expected, ArrayUtils.reverse(ArrayUtils.in1d(ar1, ar2))));
+
+		ar1 = new int[] { 2, 6, 4 };
+		expected = new int[0];
+		assertTrue(Arrays.equals(expected, ArrayUtils.in1d(ar1, ar2)));
+
+		// Test none in the second
+		assertTrue(Arrays.equals(expected, ArrayUtils.in1d(ar1, expected)));
+		// Test none in both
+		assertTrue(Arrays.equals(expected, ArrayUtils.in1d(expected, expected)));
+	}
+
+	@Test
+	public void testRecursiveCoordinatesAssemble() throws InterruptedException {
+		/*Create huge 5 dimensional matrix*/
+		int dimSize = 14, dimNumber = 5;
+		int[] dimCoordinates = new int[dimSize];
+		List<int[]> dimensions = new ArrayList<int[]>();
+		for (int i = 0; i < dimNumber; i++) {
+			for (int j = 0; j < dimSize; j++) {
+				dimCoordinates[j] = j;
+			}
+			dimensions.add(dimCoordinates);
+		}
+		long startTime = System.currentTimeMillis();
+
+		List<int[]> neighborList = ArrayUtils.dimensionsToCoordinateList(dimensions);
+		long take = System.currentTimeMillis() - startTime;
+		System.out.print("Execute in:" + take + " milliseconds");
+
+		assertEquals(neighborList.size(), 537824);
+	}
 
 	/**
 	 * Python does modulus operations differently than the rest of the world
@@ -216,7 +216,7 @@ public class ArrayUtilsTest {
 		int a = -7;
 		int n = 5;
 		assertEquals(3, ArrayUtils.modulo(a, n));
-		
+
 		//Example A
 		a = 5;
 		n = 2;
@@ -252,35 +252,35 @@ public class ArrayUtilsTest {
 		n = 10;
 		assertEquals(2, ArrayUtils.modulo(a, n));
 	}
-	
+
 	@Test
 	public void testAnd() {
 		int[] a = new int[] { 0, 0, 0, 0, 1, 1, 1 };
 		int[] b = new int[] { 0, 0, 0, 0, 1, 1, 1 };
 		int[] result = ArrayUtils.and(a, b);
 		assertTrue(Arrays.equals(a , result));
-		
+
 		a = new int[] { 0, 0, 0, 0, 1, 0, 1 };
 		result = ArrayUtils.and(a, b);
 		assertTrue(Arrays.equals(a, result));
-		
+
 		a = new int[] { 0, 0, 0, 0, 0, 0, 0 };
 		result = ArrayUtils.and(a, b);
 		assertTrue(Arrays.equals(a, result));
-		
+
 		a = new int[] { 1, 1, 1, 1, 0, 0, 0 };
 		int[] expected = new int[] { 0, 0, 0, 0, 0, 0, 0 };
 		result = ArrayUtils.and(a, b);
 		assertTrue(Arrays.equals(expected, result));
 	}
-	
+
 	@Test
 	public void testBitsToString() {
 		String expected = "c....***";
 		String result = ArrayUtils.bitsToString(new int[] { 0, 0, 0, 0, 1, 1, 1 });
 		assertEquals(expected, result);
 	}
-	
+
 	@Test
 	public void testDiff() {
 		double[] t = new double[] { 5, 4, 3, 2, 1, 0 };
@@ -290,37 +290,43 @@ public class ArrayUtilsTest {
 		assertEquals(-5, ArrayUtils.sum(result), 0);
 	}
 
-    @Test
-   	public void testMultiDimensionArrayOperation() {
-        int[] dimensions = {5, 5 ,5};
-          Object multiDimArray = createMultiDimensionArray(dimensions);
-          ArrayUtils.fillArray(multiDimArray, 1);
-          assertEquals(125, ArrayUtils.aggregateArray(multiDimArray));
-    }
+	@Test
+	public void testMultiDimensionArrayOperation() {
+		int[] dimensions = {5, 5 ,5};
+		Object multiDimArray = createMultiDimensionArray(dimensions);
+		ArrayUtils.fillArray(multiDimArray, 1);
+		assertEquals(125, ArrayUtils.aggregateArray(multiDimArray));
+	}
 
-    private Object createMultiDimensionArray(int[] sizes){
-        return Array.newInstance(int.class, sizes);
-    }
+	private Object createMultiDimensionArray(int[] sizes){
+		return Array.newInstance(int.class, sizes);
+	}
 
 	@Test
 	public void testConcatAll() {
 		assertTrue(Arrays.equals(new int[]{1, 2, 3, 4, 5, 6, 7, 8, 9, 0},
-			ArrayUtils.concatAll(new int[]{1, 2}, new int[]{3, 4, 5, 6, 7}, new int[]{8, 9, 0})));
+					ArrayUtils.concatAll(new int[]{1, 2}, new int[]{3, 4, 5, 6, 7}, new int[]{8, 9, 0})));
 	}
-	
+
 	@Test
 	public void testReplace() {
-	    assertTrue(Arrays.equals(new int[]{1, 2, 3, 4, 5, 6, 7, 8, 9, 0},
-	        ArrayUtils.replace(5, 10, new int[]{1, 2, 3, 4, 5, -1, -1, -1, -1, -1}, new int[] { 6, 7, 8, 9, 0})));
-	        
+		assertTrue(Arrays.equals(new int[]{1, 2, 3, 4, 5, 6, 7, 8, 9, 0},
+					ArrayUtils.replace(5, 10, new int[]{1, 2, 3, 4, 5, -1, -1, -1, -1, -1}, new int[] { 6, 7, 8, 9, 0})));
+
 	}
-	
+
 	@Test
 	public void testIsSparse() {
-	    int[] t = new int[] { 0, 1, 0 };
-	    int[] t1 = new int[] { 4, 5, 6, 7 };
-	    
-	    assertFalse(ArrayUtils.isSparse(t));
-	    assertTrue(ArrayUtils.isSparse(t1));
+		int[] t = new int[] { 0, 1, 0 };
+		int[] t1 = new int[] { 4, 5, 6, 7 };
+
+		assertFalse(ArrayUtils.isSparse(t));
+		assertTrue(ArrayUtils.isSparse(t1));
+	}
+
+	@Test
+	public void testNGreatest() {
+		double[] overlaps = new double[] { 1, 2, 1, 4, 8, 3, 12, 5, 4, 1 };
+		assertTrue(Arrays.equals(new int[] { 6, 4, 7 }, ArrayUtils.nGreatest(overlaps, 3)));
 	}
 }