Fixes #354 

Here I am using streams. This is not a replacement for nupic python argpartition function. I have measured the performance gain with this change. It gives atleast 4000% performance gain. Please review