@cogmission 2015-12-05
Hi @smadasu !

Sorry for the delayed response, I was away from my computer yesterday eve. I am anxious to see the difference!

The community has a contribution process which I can help you out with here. The 3 "x"'s above have to do with this process; and we must satisfy these measures before we can consider merging your contributions.

1. CL (Contributor's License) - Please fill out the form [at this link.](http://numenta.org/licenses/cl/) This will then be processed by Numenta (@rhyolight might not be able to get to this until Monday, so I appreciate your patience on this due to it being the weekend). But while you're waiting, you can... 

2. reference the "issue" your PR solves by editing your description above to include the words 
```
Fixes #354
```
on a separate line at the top... This will fix the 2nd issue entitled, "Fixes Issue Validator". All PR's in Numenta.org must be associated with an "Issue" for tracking.

3. We like for all code to be covered by tests or at least have the coverage remain constant or preferably increase. We enforce that with this 3rd requirement which does a scan of the ratio of tested lines of code to the total count of lines using the Coveralls tool. So we need you to add a test to your code in htm.java/src/test/java/org/numenta/nupic/algorithms/SpatialPoolerTest.java. 

Cheers @smadasu and thank you for your hard work!

David