@smadasu 2015-12-05
Hi David,
I have already submitted the CL yesterday. May be it will be cleared on Monday. I have updated PR description. As far as the code coverage, I have not added any new code. There was already a test covering the method I changed. Where I have taken out the need for ArrayUtils class, since it is not necessary. Which I think is the reason for code coverage going down