@cogmission 2015-12-05
Hi Sreenath,

Take a look at your comment now? ;-)  That is what I meant...

Re: [No new Code but Coveralls complains] - Sorry dude, we must make Coveralls Coverage stats always go up, so you will have to submit another test of something that is not already tested. I know its a pain but we _all_ have been through this! Lol! 

Try going to the coveralls website by clicking the "Details" link where the line "X coverage/coveralls - Coveralls decreased..." is. Then once there, navigate to the most "uncovered code and add some tests" (That will make it easy to do).

Cheers,
David