@cogmission 2015-12-07
Hi@smadasu,

 A couple of things while we wait for you CL. Can you make sure you are using spaces instead of tabs, and that the indent is set to 4 spaces instead of 8? Also, the title of this PR should keep the descriptive title, you only have to put "Fixes ###" in your comments not the title.

Once you fix the indentation and make sure you have no tabs, then everything else looks good :+1: