@smadasu 2015-12-07
Fixes #354. I have updated the spaces.