@cogmission 2015-12-07
@rhyolight please do not merge this. Just approve the CL. I need to vet the speed increase.