@cogmission 2015-12-11
Hi @smadasu,

I will look at it today. I'm sorry I've been unusually swamped with work while at the same time under the weather due to weird sleep patterns. Usually this stuff has a much faster turn around... 

Cheers,
David