@cogmission 2015-12-12
Below you can see the **enormous** difference in performance. I omit 18 of 20 iterations for brevity's sake.

Before: (SP with Global Inhibition)
```
# Run progress: 30.00% complete, ETA 00:17:02
# Fork: 10 of 10
# Warmup Iteration   1: SLF4J: Failed to load class "org.slf4j.impl.StaticLoggerBinder".
SLF4J: Defaulting to no-operation (NOP) logger implementation
SLF4J: See http://www.slf4j.org/codes.html#StaticLoggerBinder for further details.
8.764 ms/op
# Warmup Iteration  19: 7.826 ms/op
# Warmup Iteration  20: 7.809 ms/op
Iteration  19: 7.760 ms/op
Iteration  20: 7.806 ms/op
```

With @smadasu 's fix: (SP with Global Inhibition)
```
# Run progress: 30.00% complete, ETA 00:17:03
# Fork: 10 of 10
# Warmup Iteration   1: SLF4J: Failed to load class "org.slf4j.impl.StaticLoggerBinder".
SLF4J: Defaulting to no-operation (NOP) logger implementation
SLF4J: See http://www.slf4j.org/codes.html#StaticLoggerBinder for further details.
5.033 ms/op
# Warmup Iteration  19: 3.795 ms/op
# Warmup Iteration  20: 3.789 ms/op
Iteration  19: 3.835 ms/op
Iteration  20: 3.834 ms/op
```