@smadasu 2015-12-13
Fantastic @cogmission !!! I am glad that you could verify the performance boost. Thank you very much for your help. Looking forward to working with htm.java for many more fixes and enhancements. :thumbsup::smile: