@cogmission 2016-05-30
@smadasu Why the change? I'm just curious if there's something that you seen needs fixing that would help improve HTM.java? I'm open if you want to talk?