@smadasu 2016-05-31
Its just that by using JDK8 streams API, we can boost performance. Just wanted to show case it that's all :)