@cogmission 2016-06-06
@smadasu I was saying that so I could get _your_ email and write you about something...?